web-server/requirements.txt
Flask
Flask-SocketIO
nats-py
